+++ library/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the library will be documented in this file.
 - Add `toBigint`, `toBoolean`, `toDate`, `toNumber` and `toString` transformation actions (pull request #1212)
 - Add `examples` action to add example values to a schema (pull request #1199)
 - Add `getExamples` method to extract example values from a schema (pull request #1199)
+- Add `isbn` validation action to validate ISBN-10 and ISBN-13 strings (pull request #1097)
 - Add exports for `RawCheckAddIssue`, `RawCheckContext`, `RawCheckIssueInfo`, `RawTransformAddIssue`, `RawTransformContext` and `RawTransformIssueInfo` types for better developer experience with `rawCheck` and `rawTransform` actions (pull request #1359)
 - Change build step to tsdown